refactor: upload utils
